Extract getBuilt into Docker and Native strategy classes
